+++ src/terminal/InputGenerator.h
@@ -13,6 +13,8 @@
  */
 #pragma once
 
+#include <terminal/DECTextLocator.h>
+#include <terminal/Sequencer.h>
 #include <terminal/primitives.h>
 
 #include <crispy/escape.h>
@@ -47,6 +49,7 @@ enum class MouseProtocol
     AnyEventTracking = 1003,
 };
 
+// {{{ Modifier
 class Modifier
 {
   public:
@@ -214,34 +217,6 @@ enum class KeyMode
     Application
 };
 // }}}
-// {{{ Mouse
-enum class MouseButton
-{
-    Left,
-    Right,
-    Middle,
-    Release, // Button was released and/or no button is pressed.
-    WheelUp,
-    WheelDown,
-};
-
-std::string to_string(MouseButton _button);
-
-enum class MouseTransport
-{
-    // CSI M Cb Cx Cy, with Cb, Cx, Cy incremented by 0x20
-    Default,
-    // CSI M Cb Coords, with Coords being UTF-8 encoded, Coords is a tuple, each value incremented by 0x20.
-    Extended,
-    // `CSI Cb Cx Cy M` and `CSI Cb Cx Cy m` (button release)
-    SGR,
-    // SGR-Pixels (1016), an xterm extension as of Patch #359 - 2020/08/17
-    // This is just like SGR but reports pixels isntead of ANSI cursor positions.
-    SGRPixels,
-    // `CSI < Cb Cx Cy M` with Cb += 0x20
-    URXVT,
-};
-// }}}
 
 class InputGenerator
 {
@@ -337,6 +312,8 @@ class InputGenerator
     /// Resets the input generator's state, as required by the RIS (hard reset) VT sequence.
     void reset();
 
+    DECTextLocator& textLocator() noexcept { return textLocator_; }
+
   private:
     bool generateMouse(MouseEventType _eventType,
                        Modifier _modifier,
@@ -361,6 +338,14 @@ class InputGenerator
     inline bool append(uint8_t _byte);
     inline bool append(unsigned int _asciiChar);
 
+    template <typename T>
+    void executeTextLocator(T callback)
+    {
+        callback(textLocator_);
+        if (auto reply = textLocator_.peekLocatorReply(); !reply.empty())
+            generateRaw(reply);
+    }
+
     // private fields
     //
     KeyMode cursorKeysMode_ = KeyMode::Normal;
@@ -375,6 +360,8 @@ class InputGenerator
 
     std::set<MouseButton> currentlyPressedMouseButtons_ {};
     CellLocation currentMousePosition_ {}; // current mouse position
+
+    DECTextLocator textLocator_;
 };
 
 inline std::string to_string(InputGenerator::MouseEventType _value)
@@ -606,4 +593,5 @@ struct formatter<terminal::Key>
         return fmt::format_to(_ctx.out(), "{}", (unsigned) _value);
     }
 };
+
 } // namespace fmt